优化公告逻辑,修复公告样式问题
