+++ application/admin/controller/Index.php
@@ -9,6 +9,7 @@ namespace app\admin\controller;
 
 use auth\Auth;
 use app\admin\model\User;
+use app\admin\model\Config;
 
 class Index extends Common
 {
@@ -28,6 +29,9 @@ class Index extends Common
         $info = User::get($this->uid)->hidden(['password']);
         $info['head'] ? : $info['head'] = '/images/face.jpg';
         $this->assign('info', $info);
+        //公告
+        $notice_config = $this->noticeConfig();
+        $this->assign('notice_config', $notice_config);
         return $this->fetch();
     }
 
@@ -41,4 +45,21 @@ class Index extends Common
         return $this->fetch();
     }
 
+    /**
+     * 公告配置信息
+     * @return array|mixed|\PDOStatement|string|\think\Model|null
+     * @throws \think\db\exception\DataNotFoundException
+     * @throws \think\db\exception\ModelNotFoundException
+     * @throws \think\exception\DbException
+     */
+    private function noticeConfig()
+    {
+        $notice_config = cache('notice_config');
+        if ($notice_config) {
+            return $notice_config;
+        }
+        $list = Config::where('name', '=', 'notice_config')->field('value')->find();
+        cache('notice_config', $list);
+        return $list;
+    }
 }
